refactor(onboard): move landing sandbox routes
