+++ src/treg/api.py
@@ -107,12 +107,22 @@ from .routers import admin as admin_routes
 from .routers import onboard as onboard_routes
 from .routers.onboard import (
     OnboardIn,
+    SANDBOX_HIT_NS,
+    SANDBOX_RATE_MAX,
+    SANDBOX_RATE_WINDOW_S,
     TeammateIn,
+    demo_sandbox_live,
+    demo_sandbox_mint,
+    demo_sandbox_skill,
+    landing_stripe_feed,
     onboard_accept_teammate,
     onboard_demo,
     onboard_reset,
     onboard_seed_tool,
     onboard_skip,
+    skill_install,
+    skill_samples,
+    stripe_webhook,
 )
 from .routers.admin import (
     BoolIn,
@@ -880,14 +890,6 @@ router.routes.extend(org_routes.invite_entry_router.routes)
 router.routes.extend(onboard_routes.onboard_entry_router.routes)
 
 
-# ---- landing-page sandbox studio: an anonymous, throwaway team the visitor builds ----------
-# Per-IP limiter for the unauthenticated mint endpoint, in the DB (treg.ratestore) so it survives a
-# restart and holds across instances (backlog #3). It caps DB churn from the public landing page (abuse
-# is otherwise structurally contained — sandbox calls never touch the network, each sandbox is capped + TTL'd).
-SANDBOX_HIT_NS = "sandbox_hit"
-SANDBOX_RATE_MAX = 12          # sandboxes per IP per window
-SANDBOX_RATE_WINDOW_S = 3600   # 1 hour
-
 # Per-IP limiter for /call with a PUBLIC-DEMO token (the landing page publishes one shared member
 # token, so the per-user daily cap is meaningless there — thousands of strangers are one "user").
 PUBLIC_DEMO_HIT_NS = "pubdemo_call"
@@ -941,101 +943,7 @@ async def _enforce_daily_cap(caller: Caller, db: AsyncSession) -> None:
 
 
 
-@app.post("/demo/sandbox")
-async def demo_sandbox_mint(request: Request, db: AsyncSession = Depends(get_session)) -> dict:
-    """Mint a login-free, short-lived sandbox TEAM for the landing-page studio: a throwaway org + a
-    starter secret + a starter endpoint + a member token, returned so the browser (and the visitor's
-    terminal) can register more, call them, and export a skill — all with no account. Sandbox calls
-    never touch the network (see call_tool → sandbox.synthesize); rate-limited per IP; GC'd after the
-    TTL. No auth — this is the anonymous front door."""
-    await ratestore.sweep(db, SANDBOX_HIT_NS)  # evict cold IP keys so the namespace can't grow unbounded
-    if not await ratestore.rate_check(db, SANDBOX_HIT_NS,
-                                      [(_client_ip(request), SANDBOX_RATE_MAX)], SANDBOX_RATE_WINDOW_S):
-        await db.commit()  # persist the sweep even on reject
-        raise HTTPException(status_code=429, detail="too many demo sandboxes from here — try again later")
-    await db.commit()  # persist the recorded hit before minting
-    await demo_sandbox.gc(db)  # opportunistic reap of expired sandboxes
-    out = await demo_sandbox.mint(db)
-    out["live"] = bool(get_settings().demo_stripe_key)  # is the seeded stripe tool a real wire?
-    return out
-
-
-@app.get("/demo/sandbox/live")
-async def demo_sandbox_live(caller: Caller = Depends(require_member)) -> dict:
-    """Live-wire facts for an EXISTING sandbox (the browser reuses one via localStorage, so it may
-    predate the mint response carrying them): is the wire on, and who am I in the feed."""
-    if not demo_sandbox.is_sandbox(caller.org):
-        raise HTTPException(status_code=400, detail="live-wire info is for the landing-page sandbox only")
-    return {"live": bool(get_settings().demo_stripe_key),
-            "visitor": demo_sandbox.visitor_name(caller.org.slug)}
-
-
-# ---- landing-page live payments feed (the public Stripe demo — see pubfeed.py) --------------
-@app.post("/stripe/webhook", include_in_schema=False)
-async def stripe_webhook(request: Request) -> dict:
-    """Stripe → treg: a signed event from the demo sandbox account. Only `charge.succeeded` feeds
-    the landing ticker; everything else is acknowledged and dropped. 404 when unconfigured, so a
-    deploy without the secret exposes no unauthenticated POST surface."""
-    secret = get_settings().demo_stripe_webhook_secret
-    if not secret:
-        raise HTTPException(status_code=404)
-    payload = await request.body()
-    if not pubfeed.verify_signature(payload, request.headers.get("stripe-signature", ""), secret):
-        raise HTTPException(status_code=400, detail="bad signature")
-    try:
-        event = json.loads(payload)
-    except ValueError:
-        raise HTTPException(status_code=400, detail="bad payload")
-    if event.get("type") == "charge.succeeded":
-        pubfeed.push_charge(event.get("data", {}).get("object", {}) or {})
-    return {"received": True}
-
-
-@app.get("/landing/stripe-feed", include_in_schema=False)
-async def landing_stripe_feed() -> StreamingResponse:
-    """SSE stream for the landing demo pane: recent charges, then live ones. Unauthenticated by
-    design — it carries only server-chosen fields (amount/currency/created/id-suffix)."""
-    return StreamingResponse(pubfeed.stream(), media_type="text/event-stream", headers={
-        "Cache-Control": "no-cache",
-        "X-Accel-Buffering": "no",  # tell the reverse proxy not to buffer the stream
-    })
-
-
-@app.get("/demo/sandbox/skill")
-async def demo_sandbox_skill(
-    caller: Caller = Depends(require_member), db: AsyncSession = Depends(get_session)
-) -> dict:
-    """Export whatever the visitor built in their sandbox as a shareable **skill** (treg.json manifest
-    + SKILL.md + install commands). Sandbox-only — the payoff that shows what skills are."""
-    if not demo_sandbox.is_sandbox(caller.org):
-        raise HTTPException(status_code=400, detail="skill export is for the landing-page sandbox only")
-    return await demo_sandbox.export_skill(db, caller.org)
-
-
-@app.get("/skills/samples")
-async def skill_samples() -> list[dict]:
-    """The hosted sample skills the landing offers — each with its files (SKILL.md/treg.json/.secret)
-    and the prompt to try. Public: the landing renders these as file packages."""
-    base = get_settings().public_url.rstrip("/")
-    return [{"name": n, "label": s["label"], "key": s["key"], "prompt": s["prompt"],
-             "files": demo_sandbox.skill_files(n, base, None)}
-            for n, s in demo_sandbox.SAMPLE_SKILLS.items()]
-
-
-@app.get("/skills/{name}/install.sh", include_in_schema=False)
-async def skill_install(name: str, token: str = ""):
-    """`curl -fsSL {BASE}/skills/<name>/install.sh?token=<t> | sh` — writes the skill into
-    ./.claude/skills/<name>/ so Claude Code loads it. The token (if given) is baked into the
-    recipe's calls; without it the recipe reads the token from `treg login`."""
-    if name not in demo_sandbox.SAMPLE_SKILLS:
-        raise HTTPException(status_code=404, detail=f"unknown skill {name!r}")
-    # The token is interpolated into a shell script the visitor runs (`curl … | sh`). Restrict it to a
-    # real token charset so a crafted value can't inject a newline + commands into the generated script.
-    if token and not re.fullmatch(r"[A-Za-z0-9_\-]{1,200}", token):
-        raise HTTPException(status_code=422, detail="invalid token")
-    base = get_settings().public_url.rstrip("/")
-    script = demo_sandbox.install_script(name, base, token or None)
-    return PlainTextResponse(script, media_type="text/plain; charset=utf-8")
+router.routes.extend(onboard_routes.sandbox_router.routes)
 
 
 router.routes.extend(onboard_routes.onboard_teammate_router.routes)
